+++ core/txpool/legacypool/lending_pool_test.go
@@ -1,4 +1,4 @@
-package txpool
+package legacypool
 
 import (
 	"math/big"
