+++ eth/protocol_test.go
@@ -23,6 +23,7 @@ import (
 	"time"
 
 	"github.com/XinFinOrg/XDPoSChain/common"
+	"github.com/XinFinOrg/XDPoSChain/core/txpool"
 	"github.com/XinFinOrg/XDPoSChain/core/types"
 	"github.com/XinFinOrg/XDPoSChain/crypto"
 	"github.com/XinFinOrg/XDPoSChain/eth/downloader"
@@ -131,11 +132,12 @@ func testSendTransactions(t *testing.T, protocol int) {
 
 	// Fill the pool with big transactions.
 	const txsize = txsyncPackSize / 10
-	alltxs := make([]*types.Transaction, 100)
+	alltxs := make([]*txpool.Transaction, 100)
 	for nonce := range alltxs {
-		alltxs[nonce] = newTestTransaction(testAccount, uint64(nonce), txsize)
+		tx := newTestTransaction(testAccount, uint64(nonce), txsize)
+		alltxs[nonce] = &txpool.Transaction{Tx: tx}
 	}
-	pm.txpool.AddRemotes(alltxs)
+	pm.txpool.Add(alltxs, false, false)
 
 	// Connect several peers. They should all receive the pending transactions.
 	var wg sync.WaitGroup
@@ -143,7 +145,7 @@ func testSendTransactions(t *testing.T, protocol int) {
 		defer p.close()
 		seen := make(map[common.Hash]bool)
 		for _, tx := range alltxs {
-			seen[tx.Hash()] = false
+			seen[tx.Tx.Hash()] = false
 		}
 		for n := 0; n < len(alltxs) && !t.Failed(); {
 			var txs []*types.Transaction
